+++ geosongpu_ci/tools/py_ftn_interface/templates/interface.c.jinja2
@@ -31,3 +31,27 @@
 
 {% endfor %}
 
+{% for validation in validations -%}
+
+  void v_{{prefix}}_{{validation.candidate}}_c(
+      // inputs-outputs
+      {% for inout in validation.inouts -%}
+        {{inout.type}} f90_{{inout.name}}, {{inout.type}} py_{{inout.name}}{{ ", " if not loop.last or validation.outputs|length > 0 else "" }}
+      {%- endfor %}
+      // outputs
+      {% for output in validation.outputs -%}
+        {{output.type}} f90_{{output.name}}, {{output.type}} py_{{output.name}}{{ ", " if not loop.last else "" }}
+      {%- endfor %}
+  ) {
+    v_{{prefix}}_{{validation.candidate}}_py(
+      {% for inout in validation.inouts -%}
+        f90_{{inout.name}}, py_{{inout.name}}{{ ", " if not loop.last or validation.outputs|length > 0 else "" }}
+      {%- endfor %}
+      // outputs
+      {% for output in validation.outputs -%}
+        f90_{{output.name}}, py_{{output.name}}{{ ", " if not loop.last else "" }}
+      {%- endfor %}
+    );
+  }
+
+{% endfor %}
